Clear AUDC1/2 when SOUNDR is 1
If SIO sounds are enabled CH4 will output the SIO beeps and all other
channels are muted, like the OS does.

If SIO sounds are disabled CH4 will be muted but AUDC1/2 aren't touched
so that sound can be played on channels 1/2.

Also clear AUDC3 in both cases to avoid issues after switching back to
8bit mode.

Signed-off-by: Matthias Reichl <hias@horus.com>