Fail the build if critical code crosses a page boundary
If the branches in GETLP and VBICODE cross a page boundary receiving
at 126kbit can become unreliable as the code takes too many cycles to
execute.

While failing the build in release mode is desirable it can be annoying
during development so the error can be turned into a warning by defining
the ALLOW_PAGE_CROSSING symbol.

Signed-off-by: Matthias Reichl <hias@horus.com>